backend/requirements.txt
beautifulsoup4==4.13.3
fastapi==0.115.11
mysql_connector_repackaged==0.3.1
pydantic==2.10.6
pytest==8.3.5
Requests==2.32.3
spotipy==2.25.0
pluggy
httpx
mysql-connector-python==9.2.0
uvicorn
loguru
prometheus-client
